bigint

+++ db/scripts/003_ddl_create_car_table.sql
@@ -1,7 +1,7 @@
 create table car (
     id       bigserial primary key,
     name     varchar   not null,
-    mark_id  int       not null references mark(id),
+    mark_id  bigint    not null references mark(id),
     owners   varchar,
     color    varchar(100) not null
 );
