+++ db/scripts/004_ddl_create_auto_post_table.sql
@@ -3,6 +3,6 @@ create table auto_post (
     description    text not null,
     created        timestamp not null,
     sold           boolean not null,
-    car_id         int references car(id) unique not null,
-    user_id        int references auto_user(id) not null
+    car_id         bigint references car(id) unique not null,
+    user_id        bigint references auto_user(id) not null
 );
