+++ db/scripts/006_ddl_create_participates_table.sql
@@ -1,7 +1,7 @@
 create table participates (
     id      bigserial primary key,
-    user_id int not null,
-    post_id int not null,
+    user_id bigint not null,
+    post_id bigint not null,
         unique (user_id, post_id),
     foreign key (user_id) references auto_user(id) on delete cascade,
     foreign key (post_id) references auto_post(id) on delete cascade
