+++ db/scripts/007_ddl_create_files_table.sql
@@ -1,6 +1,6 @@
 create table files (
-    id   serial primary key,
-    name varchar not null,
-    path varchar not null unique,
-    post_id int references auto_post(id) not null
+    id      serial primary key,
+    name    varchar not null,
+    path    varchar not null unique,
+    post_id bigint references auto_post(id) not null
 );
